moved col_delete into journalist_app/utils.py
